implement new span management

+++ rpmalloc/rpmalloc.h
@@ -69,15 +69,17 @@ typedef struct rpmalloc_config_t {
 	//  will be passed to the memory unmap function. The alignment offset MUST NOT be
 	//  larger than 65535 (storable in an uint16_t), if it is you must use natural
 	//  alignment to shift it into 16 bits. If you set a memory_map function, you
-	//  must also set a memory_unmap function (or else default implementation will be used)
+	//  must also set a memory_unmap function or else the default implementation will
+	//  be used for both.
 	void* (*memory_map)(size_t size, size_t* offset);
 	//! Unmap the memory pages starting at address and spanning the given number of bytes.
-	//  If release is set to 1, the unmap is for an entire span range as returned by
-	//  a previous call to memory_map and that the entire range should be released.
-	//  If release is set to 0, the unmap is a partial decommit of a subset of the mapped
-	//  memory range. If you set a memory_unmap function, you must also set a memory_map
-	//  function (or else default implementation will be used)
-	void (*memory_unmap)(void* address, size_t size, size_t offset, int release);
+	//  If release is set to non-zero, the unmap is for an entire span range as returned by
+	//  a previous call to memory_map and that the entire range should be released. The
+	//  release argument holds the size of the entire span range. If release is set to 0,
+	//  the unmap is a partial decommit of a subset of the mapped memory range.
+	//  If you set a memory_unmap function, you must also set a memory_map function or
+	//  else the default implementation will be used for both.
+	void (*memory_unmap)(void* address, size_t size, size_t offset, size_t release);
 	//! Size of memory pages. The page size MUST be a power of two in [512,65536] range
 	//  (2^9 to 2^16) unless 0 - set to 0 to use system page size. All memory mapping
 	//  requests to memory_map will be made with size set to a multiple of the page size.
